+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/nodes/op/LLVMAddressEqualsNode.java
@@ -38,14 +38,19 @@ import com.oracle.truffle.api.library.CachedLibrary;
 import com.oracle.truffle.llvm.runtime.library.internal.LLVMNativeLibrary;
 import com.oracle.truffle.llvm.runtime.nodes.api.LLVMExpressionNode;
 import com.oracle.truffle.llvm.runtime.nodes.api.LLVMNode;
+import com.oracle.truffle.llvm.runtime.nodes.util.LLVMSameObjectNode;
 import com.oracle.truffle.llvm.runtime.pointer.LLVMManagedPointer;
 import com.oracle.truffle.llvm.runtime.pointer.LLVMNativePointer;
 import com.oracle.truffle.llvm.runtime.pointer.LLVMPointer;
-import com.oracle.truffle.llvm.spi.ReferenceLibrary;
 
 @NodeChild(type = LLVMExpressionNode.class)
 @NodeChild(type = LLVMExpressionNode.class)
+@GenerateUncached
 public abstract class LLVMAddressEqualsNode extends LLVMAbstractCompareNode {
+
+    public static LLVMAddressEqualsNode create() {
+        return LLVMAddressEqualsNodeGen.create(null, null);
+    }
 
     // the first two cases are redundant but much more efficient than the ones below
 
@@ -106,10 +111,10 @@ public abstract class LLVMAddressEqualsNode extends LLVMAbstractCompareNode {
     abstract static class LLVMManagedEqualsNode extends LLVMNode {
         abstract boolean execute(Object a, Object b);
 
-        @Specialization(limit = "3")
+        @Specialization
         protected boolean doForeign(LLVMManagedPointer a, LLVMManagedPointer b,
-                        @CachedLibrary("a.getObject()") ReferenceLibrary pointToSameObject) {
-            return pointToSameObject.isSame(a.getObject(), b.getObject()) && a.getOffset() == b.getOffset();
+                        @Cached LLVMSameObjectNode pointToSameObject) {
+            return pointToSameObject.execute(a.getObject(), b.getObject()) && a.getOffset() == b.getOffset();
         }
 
         @Specialization(guards = "isNative(p1) || isNative(p2)")